Fix codacy - AdminController Refacto
